refactor : 훅에 'use client' 추가한다
