+++ packages/web/src/app/[lng]/(main)/community/components/ContentSection.client.tsx
@@ -4,7 +4,6 @@ import { useQueryClient } from '@tanstack/react-query';
 import { Suspense } from 'react';
 
 import AllContent from './AllContent.client';
-import CreateArticleButton from './CreateArticleButton';
 import KpopContent from './KpopContent';
 import LanguageContent from './LanguageContent.client';
 import QuestionContent from './QuestionContent.client';
@@ -12,7 +11,7 @@ import QuestionContent from './QuestionContent.client';
 import { Keys } from '@/apis/community';
 import { useTranslation } from '@/app/i18n/client';
 import { Tabs } from '@/components/Tabs';
-import { useBroadcastChannel } from '@/hooks/useBroadcast';
+import { useBroadcastChannel } from '@/hooks';
 
 export interface CommunityChannelMessage {
   categoryId: number;
