+++ packages/web/src/constants/dummyData.ts
@@ -1,3 +1,5 @@
+import type { MatesResponse } from '@/apis/profile';
+
 import {
   Apply,
   Comment,
@@ -6,8 +8,6 @@ import {
   GroupMember,
   Grouping,
 } from '@/apis/groups';
-
-import type { MatesResponse } from '@/apis/profile';
 
 export const DUMMY_GROUPING_DATA: Grouping[] = [
   {
