+++ packages/web/src/hooks/useAppRouter.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useRouter } from 'next/navigation';
 
 import { sendMessageToReactNative } from '@/utils';
