+++ packages/web/src/hooks/useBottomSheet.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useCallback, useState } from 'react';
 
 export default function useBottomSheet() {
