+++ packages/web/src/hooks/useBrowser.ts
@@ -1,3 +1,4 @@
+'use client';
 import { useState } from 'react';
 
 import useDidMount from './useDidMount';
