+++ packages/web/src/hooks/useControllers.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import {
   type FieldPath,
   type FieldValues,
