+++ packages/web/src/hooks/useDidMount.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect, useRef } from 'react';
 
 export default function useDidMount(callback: VoidFunction) {
