+++ packages/web/src/hooks/useDidUnMount.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect } from 'react';
 
 export default function useDidUnMount(callback: VoidFunction) {
