+++ packages/web/src/hooks/useDidUpdate.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { DependencyList, useEffect, useRef } from 'react';
 
 export default function useDidUpdate(callback: VoidFunction, dependencyList: DependencyList) {
