+++ packages/web/src/hooks/useLogin.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import useAppRouter from './useAppRouter';
 
 import { setTokenAtCookie } from '@/utils';
