+++ packages/web/src/hooks/useLogout.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useRouter } from 'next/navigation';
 
 import { removeToken } from '@/utils';
