+++ packages/web/src/hooks/useOnClickInside.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect } from 'react';
 
 export default function useOnClickInside<T extends HTMLElement = HTMLElement>(
