+++ packages/web/src/hooks/useSentry.ts
@@ -1,3 +1,5 @@
+'use client';
+
 import * as Sentry from '@sentry/nextjs';
 import { useEffect } from 'react';
